Add a new Signer API for creating simple signing signatures with Sequoia-PGP
This API is proven end-to-end in https://github.com/containers/image/pull/2876 and
https://github.com/containers/skopeo/pull/2645 , but it is not yet convenient to
use becahse the Rust dependency has to be compiled manually.

So, for now, add the API as a stub only; that allows building the CLIs
and tests on top, and they will light up once the backend is added.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>