RFC: replace docker/client with fsouza/go-dockerclient
The motivation is to reduce the resulting binary size by switching to a
less bloated docker client implementation.

Testing with skopeo I see a 1.3MB improvements and with podman it is
around 1.2 MB on my fedora laptop.

Now there is one big catch, the ExportImage() image call only accepts an
io.Writer and thus not return an io.Reader. This means we cannot wrap it
directly to the decompress stream without making another copy. To keep
things simple I just write the output to a file and then we read that
again. Not sure how important this code path is and if we care about the
performance here.

Also note github.com/fsouza/go-dockerclient is already a dependency in
podman/buildah so for them this does not add a new dep and we can
instead get rid of the client.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>